Improves the game flow in Nick's tests.
The reward for doing well on a test is to allow the player know more about Cecilia.
Nick's Test dialogs are now in two files: one for the intro, one for the questions.
Added signals so mini-games can notify if the player wins of loses.
Added Game/Random action to set a random value between 1 and 100 in the variable "random".
Minor updates and fixes in character animations.